backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
SQLAlchemy==2.0.49
alembic==1.18.4
pytest==8.3.5
pytest-flask==1.3.0

# SaaS mode dependencies (not required for single/self-hosted mode)
PyJWT>=2.8.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
